Test modifications due to implementation changes
