+++ report/custom_test.go
@@ -7,17 +7,10 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/google/go-cmp/cmp"
 	"github.com/tenntenn/golden"
-	"golang.org/x/text/language"
 )
 
 func TestCustomMetricSetTable(t *testing.T) {
-	locales := []*language.Tag{
-		nil,
-		&language.Japanese,
-		&language.French,
-	}
 	tests := []struct {
 		s *CustomMetricSet
 	}{
@@ -58,27 +51,24 @@ func TestCustomMetricSetTable(t *testing.T) {
 	t.Setenv("GITHUB_SERVER_URL", "https://github.com")
 	t.Setenv("GITHUB_REPOSITORY", "owner/repo")
 	for i, tt := range tests {
-		for _, locale := range locales {
-			SetLocale(locale)
-			var tc string
-			if locale != nil {
-				tc = fmt.Sprintf("%d.%s", i, locale.String())
-			} else {
-				tc = fmt.Sprintf("%d", i)
-			}
-
-			t.Run(tc, func(t *testing.T) {
-				got := tt.s.Table()
-				f := filepath.Join("custom_metrics", fmt.Sprintf("custom_metric_set_table.%s", tc))
-				if os.Getenv("UPDATE_GOLDEN") != "" {
-					golden.Update(t, testdataDir(t), f, got)
-					return
-				}
-				if diff := golden.Diff(t, testdataDir(t), f, got); diff != "" {
-					t.Error(diff)
-				}
-			})
+		var tc string
+		if tt.s.report != nil && tt.s.report.config.Locale != nil {
+			tc = fmt.Sprintf("%d.%s", i, tt.s.report.config.Locale.String())
+		} else {
+			tc = fmt.Sprintf("%d", i)
 		}
+
+		t.Run(tc, func(t *testing.T) {
+			got := tt.s.Table()
+			f := filepath.Join("custom_metrics", fmt.Sprintf("custom_metric_set_table.%s", tc))
+			if os.Getenv("UPDATE_GOLDEN") != "" {
+				golden.Update(t, testdataDir(t), f, got)
+				return
+			}
+			if diff := golden.Diff(t, testdataDir(t), f, got); diff != "" {
+				t.Error(diff)
+			}
+		})
 	}
 }
 
@@ -114,12 +104,6 @@ func TestCustomMetricSetMetadataTable(t *testing.T) {
 }
 
 func TestCustomMetricSetOut(t *testing.T) {
-	locales := []*language.Tag{
-		nil,
-		&language.Japanese,
-		&language.French,
-	}
-
 	tests := []struct {
 		s *CustomMetricSet
 	}{
@@ -152,30 +136,27 @@ func TestCustomMetricSetOut(t *testing.T) {
 		}},
 	}
 	for i, tt := range tests {
-		for _, locale := range locales {
-			SetLocale(locale)
-			var tc string
-			if locale != nil {
-				tc = fmt.Sprintf("%d.%s", i, locale.String())
-			} else {
-				tc = fmt.Sprintf("%d", i)
-			}
-
-			t.Run(tc, func(t *testing.T) {
-				got := new(bytes.Buffer)
-				if err := tt.s.Out(got); err != nil {
-					t.Fatal(err)
-				}
-				f := filepath.Join("custom_metrics", fmt.Sprintf("custom_metric_set_out.%s", tc))
-				if os.Getenv("UPDATE_GOLDEN") != "" {
-					golden.Update(t, testdataDir(t), f, got)
-					return
-				}
-				if diff := golden.Diff(t, testdataDir(t), f, got); diff != "" {
-					t.Error(diff)
-				}
-			})
+		var tc string
+		if tt.s.report != nil && tt.s.report.config.Locale != nil {
+			tc = fmt.Sprintf("%d.%s", i, tt.s.report.config.Locale.String())
+		} else {
+			tc = fmt.Sprintf("%d", i)
 		}
+
+		t.Run(tc, func(t *testing.T) {
+			got := new(bytes.Buffer)
+			if err := tt.s.Out(got); err != nil {
+				t.Fatal(err)
+			}
+			f := filepath.Join("custom_metrics", fmt.Sprintf("custom_metric_set_out.%s", tc))
+			if os.Getenv("UPDATE_GOLDEN") != "" {
+				golden.Update(t, testdataDir(t), f, got)
+				return
+			}
+			if diff := golden.Diff(t, testdataDir(t), f, got); diff != "" {
+				t.Error(diff)
+			}
+		})
 	}
 }
 
@@ -240,12 +221,6 @@ func TestCustomMetricsSetValidate(t *testing.T) {
 }
 
 func TestDiffCustomMetricSetTable(t *testing.T) {
-	locales := []*language.Tag{
-		nil,
-		&language.Japanese,
-		&language.French,
-	}
-
 	tests := []struct {
 		a *CustomMetricSet
 		b *CustomMetricSet
@@ -299,28 +274,25 @@ func TestDiffCustomMetricSetTable(t *testing.T) {
 	t.Setenv("GITHUB_SERVER_URL", "https://github.com")
 	t.Setenv("GITHUB_REPOSITORY", "owner/repo")
 	for i, tt := range tests {
-		for _, locale := range locales {
-			SetLocale(locale)
-			var tc string
-			if locale != nil {
-				tc = fmt.Sprintf("%d.%s", i, locale.String())
-			} else {
-				tc = fmt.Sprintf("%d", i)
-			}
-
-			t.Run(tc, func(t *testing.T) {
-				d := tt.a.Compare(tt.b)
-				got := d.Table()
-				f := filepath.Join("custom_metrics", fmt.Sprintf("diff_custom_metric_set_table.%s", tc))
-				if os.Getenv("UPDATE_GOLDEN") != "" {
-					golden.Update(t, testdataDir(t), f, got)
-					return
-				}
-				if diff := golden.Diff(t, testdataDir(t), f, got); diff != "" {
-					t.Error(diff)
-				}
-			})
+		var tc string
+		if tt.a.report != nil && tt.a.report.config.Locale != nil {
+			tc = fmt.Sprintf("%d.%s", i, tt.a.report.config.Locale.String())
+		} else {
+			tc = fmt.Sprintf("%d", i)
 		}
+
+		t.Run(tc, func(t *testing.T) {
+			d := tt.a.Compare(tt.b)
+			got := d.Table()
+			f := filepath.Join("custom_metrics", fmt.Sprintf("diff_custom_metric_set_table.%s", tc))
+			if os.Getenv("UPDATE_GOLDEN") != "" {
+				golden.Update(t, testdataDir(t), f, got)
+				return
+			}
+			if diff := golden.Diff(t, testdataDir(t), f, got); diff != "" {
+				t.Error(diff)
+			}
+		})
 	}
 }
 
@@ -387,45 +359,6 @@ func TestDiffCustomMetricSetMetadataTable(t *testing.T) {
 				return
 			}
 			if diff := golden.Diff(t, testdataDir(t), f, got); diff != "" {
-				t.Error(diff)
-			}
-		})
-	}
-}
-
-func TestConvertFormat(t *testing.T) {
-	tests := []struct {
-		n    interface{}
-		want string
-	}{
-		{
-			int(10),
-			"10",
-		},
-		{
-			int32(3200),
-			"3,200",
-		},
-		{
-			float32(10.0),
-			"10",
-		},
-		{
-			float32(1000.1),
-			"1,000.1",
-		},
-		{
-			int(1000),
-			"1,000",
-		},
-	}
-
-	SetLocale(&language.Japanese)
-
-	for i, tt := range tests {
-		t.Run(fmt.Sprintf("%d", i), func(t *testing.T) {
-			got := convertFormat(tt.n)
-			if diff := cmp.Diff(got, tt.want, nil); diff != "" {
 				t.Error(diff)
 			}
 		})
