+++ report/report_test.go
@@ -16,6 +16,7 @@ import (
 	"github.com/k1LoW/octocov/coverage"
 	"github.com/k1LoW/octocov/gh"
 	"github.com/k1LoW/octocov/ratio"
+	"golang.org/x/text/language"
 )
 
 func TestNew(t *testing.T) {
@@ -535,4 +536,44 @@ func coverageTestdataDir(t *testing.T) string {
 		t.Fatal(err)
 	}
 	return dir
+}
+
+func TestConvertFormat(t *testing.T) {
+	tests := []struct {
+		n    interface{}
+		want string
+	}{
+		{
+			int(10),
+			"10",
+		},
+		{
+			int32(3200),
+			"3,200",
+		},
+		{
+			float32(10.0),
+			"10",
+		},
+		{
+			float32(1000.1),
+			"1,000.1",
+		},
+		{
+			int(1000),
+			"1,000",
+		},
+	}
+
+	for i, tt := range tests {
+		t.Run(fmt.Sprintf("%d", i), func(t *testing.T) {
+			l := language.Japanese
+			r := &Report{config: &ReportOptions{Locale: &l}}
+
+			got := r.convertFormat(tt.n)
+			if diff := cmp.Diff(got, tt.want, nil); diff != "" {
+				t.Error(diff)
+			}
+		})
+	}
 }
